+++ src/lib/components/fantasy/FantasyCardMobile.svelte
@@ -3,24 +3,12 @@
 	import type { Tables } from "$lib/types/database.helper.types";
 	import type { FantasyForm, FullPlayer } from "$lib/types/newTypes";
 	import Card from "../cards/Card.svelte";
+	import FantasyCardButtons from "./FantasyCardButtons.svelte";
 
     export let player: FullPlayer | null;
 	export let fantasyForm: FantasyForm;
 	export let position: number;
     export let season: Tables<'seasons'> | null;
-
-    function handleCaptainClick(playerId: number) {
-		fantasyForm.captainId = playerId;
-	}
-
-	function sellPlayer() {
-		fantasyForm.players[position] = null;
-		if (fantasyForm.captainId == player?.id) {
-			fantasyForm.captainId = -1;
-		}
-	}
-
-	$: isCaptain = fantasyForm.captainId == player?.id;
 </script>
 
 {#if player}
@@ -34,30 +22,7 @@
         </div>
             
         <div class="flex flex-row space-x-2">
-            <button type="button" class="btn px-2" on:click={() => sellPlayer()}>
-                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                    <path stroke-linecap="round" stroke-linejoin="round" d="M9.75 9.75l4.5 4.5m0-4.5l-4.5 4.5M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-                </svg>                                                
-            </button>
-            <button type="button" class="btn px-2" on:click={() => handleCaptainClick(player?.id ?? -1)}>
-                {#if isCaptain}
-                    <svg class="w-6 h-6 text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 20 20">
-                        <path d="M10 .5a9.5 9.5 0 1 0 9.5 9.5A9.51 9.51 0 0 0 10 .5Zm3.707 8.207-4 4a1 1 0 0 1-1.414 0l-2-2a1 1 0 0 1 1.414-1.414L9 10.586l3.293-3.293a1 1 0 0 1 1.414 1.414Z"/>
-                    </svg>
-                {:else}
-                    <svg class="w-6 h-6 text-gray-800 dark:text-white" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
-                        <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="m7 10 2 2 4-4m6 2a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z"/>
-                    </svg>
-                {/if}
-                
-            </button>
-            <a href="/players/{player.id}" target="_blank" rel="noreferrer" class="p-0">
-                <button type="button" class="btn px-2">
-                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                        <path stroke-linecap="round" stroke-linejoin="round" d="M17.982 18.725A7.488 7.488 0 0012 15.75a7.488 7.488 0 00-5.982 2.975m11.963 0a9 9 0 10-11.963 0m11.963 0A8.966 8.966 0 0112 21a8.966 8.966 0 01-5.982-2.275M15 9.75a3 3 0 11-6 0 3 3 0 016 0z" />
-                    </svg>                      
-                </button>
-            </a>
+            <FantasyCardButtons bind:fantasyForm={fantasyForm} player={player} position={position} />
         </div>
     </div>
 {:else}
